Вынес посредников и контроллер в отдельные сущности, в разные файлы.
